--- array_to_numpy.py
@@ -1,13 +1,0 @@
-import numpy as np
-from PIL import Image
-
-#Create 3d numpy array of zeros, then repace zeros (black pixels) with yellow pixels
-data = np.zeros((5, 4, 3), dtype=np.uint8)
-data[:] = [255, 255, 0]
-print(data)
-
-#Make a red Patch
-data[0:3, 0:2] = [255, 200, 233]
-
-img = Image.fromarray(data, 'RGB')
-img.save('canvas.png')
